pyproject.toml
[tool.poetry]
name = "billing-api"
version = "0.1.0"
description = "API to demonstrate Django and DRF concepts"
authors = ["Muchai N."]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.1"
redis = "^4.3.4"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.0"
django-model-utils = "^4.2.0"
django-auditlog = "^2.1.1"
django-split-settings = "^1.1.0"
django-environ = "^0.9.0"
python-decouple = "^3.6"
django-filter = "^22.1"
django-csp = "^3.7"
pytz = "^2022.2.1"
requests = "^2.28.1"
celery = "^5.2.7"
sentry-sdk = "^1.9.8"
django-cors-headers = "^3.5.0"
bcrypt = "^4.0"
psycopg2-binary = "^2.9.3"
flower = "^1.2.0"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^3.6.0"
nplusone = "^1.0.0"
django-extensions = "^3.2.1"
django-querycount = "^0.7.0"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
factory-boy = "^3.2.1"
locust = "^2.12.1"
black = "^22.8.0"
pylint-django = "^2.5.3"
watchdog = "^2.1.9"
line-profiler = "^3.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
